feat(merge): one run control for one result
The explorer had two run buttons once a merge existed: its own, which ran the
query the merge was built from, and the merge's, which ran the merge. Both
produced a result and only one of them reached the screen, so the other was a
warehouse query spent on an answer nobody would see — and which button to
press was left to the user to work out.

Run query runs the merge when one is configured, and says so. The merge's own
button is gone, and the query it was built from stops running once the merge
has results.

The setup panel keeps what it is for — the relationship, and what is missing
before it can run. Everything derived from the two queries moves to a hook,
because the run control is no longer inside the panel and both need the same
answer to what the merge is and whether it can run.
