+++ packages/frontend/src/ee/features/aiCopilot/components/InstructionsSupport.module.css
@@ -1,0 +1,7 @@
+.chevron {
+    transition: transform 150ms ease;
+}
+
+.chevronOpen {
+    transform: rotate(180deg);
+}
