+++ packages/frontend/src/features/mergeQuery/components/MergeQueryStrip.tsx
@@ -1,17 +1,4 @@
-import {
-    FeatureFlags,
-    getItemMap,
-    isDimension,
-    MergeQueryErrorKind,
-    validateMergeQuery,
-    type Explore,
-    type MergeFieldTypes,
-    getUnaccountedDimensions,
-    isField,
-    MergeJoinType,
-    type MergeQuery,
-    type MetricQuery,
-} from '@lightdash/common';
+import { FeatureFlags, MergeJoinType } from '@lightdash/common';
 import {
     ActionIcon,
     Alert,
@@ -33,10 +20,8 @@ import {
     IconPlus,
     IconX,
 } from '@tabler/icons-react';
-import { useCallback, useEffect, useMemo, useRef, type FC } from 'react';
+import { useEffect, useRef, type FC } from 'react';
 import MantineIcon from '../../../components/common/MantineIcon';
-import { useExplore } from '../../../hooks/useExplore';
-import { useProjectUuid } from '../../../hooks/useProjectUuid';
 import { useServerFeatureFlag } from '../../../hooks/useServerOrClientFeatureFlag';
 import {
     selectMetricQuery,
@@ -44,7 +29,7 @@ import {
     useExplorerSelector,
 } from '../../explorer/store';
 import { useMergeSafe } from '../context/useMerge';
-import { useMergePivotValues } from '../hooks/useMergeQuery';
+import { useMergeSetup } from '../hooks/useMergeSetup';
 
 const MAX_PIVOT_VALUES = 50;
 
@@ -72,9 +57,6 @@ const EMPTY_MERGE = {
     setPivotValues: () => {},
     setPostPivotIndex: () => {},
 };
-const SOURCE_A = 'a';
-const SOURCE_B = 'b';
-const JOIN_KEY = 'join_key';
 
 /**
  * One query in the merge. Clicking it re-targets the field picker.
@@ -157,7 +139,6 @@ const QueryRow: FC<{
  * belongs to the join line because it describes the relationship.
  */
 export const MergeQueryStrip: FC = () => {
-    const projectUuid = useProjectUuid();
     const { data: mergeFlag } = useServerFeatureFlag(FeatureFlags.MergeQueries);
     const tableName = useExplorerSelector(selectTableName);
     const metricQuery = useExplorerSelector(selectMetricQuery);
@@ -167,9 +148,7 @@ export const MergeQueryStrip: FC = () => {
         wasRestored,
         focus,
         queryB,
-        joinParts,
         joinType,
-        pivotValues,
         postPivotIndex,
         addQuery,
         removeQuery,
@@ -182,258 +161,30 @@ export const MergeQueryStrip: FC = () => {
         setPostPivotIndex,
     } = mergeContext ?? EMPTY_MERGE;
 
-    const { run, isRunning, runErrors, mergeResults } = mergeContext ?? {};
+    const { isRunning, runErrors, mergeResults } = mergeContext ?? {};
 
-    // The first key part defaults to each query's first dimension. Picking a
-    // dimension and then picking it again as the join field is the same choice
-    // twice; further parts start empty because there is no obvious default.
-    const effectiveParts = useMemo(
-        () =>
-            joinParts.map((part, index) => ({
-                fieldA:
-                    part.fieldA ??
-                    (index === 0 ? (metricQuery.dimensions[0] ?? null) : null),
-                fieldB:
-                    part.fieldB ??
-                    (index === 0 ? (queryB.dimensions[0] ?? null) : null),
-            })),
-        [joinParts, metricQuery.dimensions, queryB.dimensions],
-    );
-    const completeParts = effectiveParts.filter(
-        (part) => part.fieldA && part.fieldB,
-    );
-
-    const { data: exploreA } = useExplore(tableName, {
-        refetchOnMount: false,
-    });
-    const { data: exploreB } = useExplore(queryB.exploreName ?? undefined, {
-        refetchOnMount: false,
-    });
-    // Field ids are how the merge is addressed, but they are not what anyone
-    // calls these things. Everything the user reads says the label.
-    const labelFor = useCallback(
-        (fieldId: string) => {
-            const item =
-                (exploreA ? getItemMap(exploreA)[fieldId] : undefined) ??
-                (exploreB ? getItemMap(exploreB)[fieldId] : undefined);
-            return item && isField(item) ? item.label : fieldId;
-        },
-        [exploreA, exploreB],
-    );
-
-    const metricQueryB = useMemo<MetricQuery>(
-        () => ({
-            exploreName: queryB.exploreName ?? '',
-            dimensions: queryB.dimensions,
-            metrics: queryB.metrics,
-            filters: {},
-            sorts: [],
-            limit: metricQuery.limit,
-            tableCalculations: [],
-        }),
-        [queryB, metricQuery.limit],
-    );
-
-    // Either query can be the finer-grained one. Both are checked, because a
-    // merge is refused for whichever side carries the extra dimension and the
-    // repair has to be offered where the problem is.
-    const unaccountedA = useMemo(
-        () =>
-            getUnaccountedDimensions(
-                { id: SOURCE_A, metricQuery, pivot: null },
-                completeParts.map((part, index) => ({
-                    name: `${JOIN_KEY}_${index}`,
-                    fieldIdBySourceId: { [SOURCE_A]: part.fieldA as string },
-                })),
-            ),
-        [metricQuery, completeParts],
-    );
-    const unaccountedB = useMemo(
-        () =>
-            getUnaccountedDimensions(
-                { id: SOURCE_B, metricQuery: metricQueryB, pivot: null },
-                completeParts.map((part, index) => ({
-                    name: `${JOIN_KEY}_${index}`,
-                    fieldIdBySourceId: { [SOURCE_B]: part.fieldB as string },
-                })),
-            ),
-        [metricQueryB, completeParts],
-    );
-
-    // One side at a time: repairing both at once needs a value set per side,
-    // which the setup does not yet ask for.
-    const pivotSide: 'a' | 'b' | null =
-        unaccountedA.length === 1 && unaccountedB.length === 0
-            ? 'a'
-            : unaccountedB.length === 1 && unaccountedA.length === 0
-              ? 'b'
-              : null;
-    const unaccounted = pivotSide === 'b' ? unaccountedB : unaccountedA;
-    const pivotField = pivotSide ? unaccounted[0] : null;
-    const pivotQueryLabel = pivotSide === 'b' ? 'Query B' : 'Query A';
-    const otherQueryLabel = pivotSide === 'b' ? 'Query A' : 'Query B';
-    const pivotFieldLabel = pivotField ? labelFor(pivotField) : '';
-    const joinFieldLabel = effectiveParts[0]?.fieldA
-        ? labelFor(effectiveParts[0].fieldA)
-        : 'the join key';
-
-    const { data: pivotValueOptions, isLoading: isLoadingValues } =
-        useMergePivotValues(
-            projectUuid,
-            pivotSide === 'b' ? metricQueryB : metricQuery,
-            pivotField,
-            MAX_PIVOT_VALUES,
-        );
-    const suggestedValues = pivotValueOptions?.values ?? [];
-    const effectivePivotValues =
-        pivotValues.length > 0 ? pivotValues : suggestedValues;
-
-    const unaccountedTotal = unaccountedA.length + unaccountedB.length;
-    // Built here rather than inside the run handler so the same object can be
-    // validated while it is being configured. The rules that refuse a merge do
-    // not need it to have run.
-    const mergeQuery = useMemo<MergeQuery | null>(() => {
-        if (!queryB.exploreName || completeParts.length === 0) return null;
-
-        const joinKey = completeParts.map((part, index) => ({
-            name: `${JOIN_KEY}_${index}`,
-            fieldIdBySourceId: {
-                [SOURCE_A]: part.fieldA as string,
-                [SOURCE_B]: part.fieldB as string,
-            },
-        }));
-
-        return {
-            sources: [
-                {
-                    id: SOURCE_A,
-                    metricQuery,
-                    pivot:
-                        pivotSide === 'a' && pivotField
-                            ? {
-                                  fieldId: pivotField,
-                                  values: effectivePivotValues,
-                                  includeNulls: false,
-                              }
-                            : null,
-                },
-                {
-                    id: SOURCE_B,
-                    metricQuery: metricQueryB,
-                    pivot:
-                        pivotSide === 'b' && pivotField
-                            ? {
-                                  fieldId: pivotField,
-                                  values: effectivePivotValues,
-                                  includeNulls: false,
-                              }
-                            : null,
-                },
-            ],
-            joinKey,
-            joinType,
-            postPivot:
-                postPivotIndex !== null && joinKey[postPivotIndex]
-                    ? {
-                          keyName: joinKey[postPivotIndex].name,
-                          values: effectivePivotValues,
-                          includeNulls: false,
-                      }
-                    : null,
-            tableCalculations: [],
-            limit: metricQuery.limit,
-        };
-    }, [
-        queryB,
-        completeParts,
-        metricQuery,
-        metricQueryB,
+    const {
+        effectiveParts,
+        labelFor,
+        unaccountedA,
+        unaccountedB,
+        unaccountedTotal,
         pivotSide,
         pivotField,
+        pivotFieldLabel,
+        pivotQueryLabel,
+        otherQueryLabel,
+        joinFieldLabel,
+        suggestedValues,
+        isLoadingValues,
+        pivotValueOptions,
         effectivePivotValues,
-        joinType,
-        postPivotIndex,
-    ]);
-
-    // The same rules the server refuses on, run here as the merge is built.
-    // Whether two fields can be joined depends only on the two fields, so
-    // making the user press Run to find out is a round trip for an answer we
-    // already have.
-    const joinFieldTypes = useMemo<MergeFieldTypes>(() => {
-        const collect = (explore: Explore | undefined) =>
-            explore
-                ? Object.entries(getItemMap(explore)).flatMap(([id, item]) =>
-                      isDimension(item)
-                          ? [
-                                [
-                                    id,
-                                    {
-                                        type: item.type,
-                                        timeInterval: item.timeInterval ?? null,
-                                    },
-                                ] as const,
-                            ]
-                          : [],
-                  )
-                : [];
-        return Object.fromEntries([...collect(exploreA), ...collect(exploreB)]);
-    }, [exploreA, exploreB]);
-
-    const joinKeyErrors = useMemo(
-        () =>
-            mergeQuery
-                ? validateMergeQuery(mergeQuery, joinFieldTypes).filter(
-                      (error) =>
-                          error.kind ===
-                              MergeQueryErrorKind.JOIN_KEY_TYPE_MISMATCH ||
-                          error.kind ===
-                              MergeQueryErrorKind.JOIN_KEY_GRANULARITY_MISMATCH,
-                  )
-                : [],
-        [mergeQuery, joinFieldTypes],
-    );
-
-    // A merge that is not built yet and a merge that is built but unsafe are
-    // different problems. Saying both at once is what makes the panel
-    // unreadable: a grain warning means nothing until there is a merge to
-    // warn about.
-    const setupStep = !queryB.exploreName
-        ? 'Pick a table for Query B'
-        : queryB.metrics.length === 0
-          ? 'Pick at least one metric for Query B'
-          : !effectiveParts.every(
-                  (part) =>
-                      part.fieldA &&
-                      part.fieldB &&
-                      metricQuery.dimensions.includes(part.fieldA) &&
-                      queryB.dimensions.includes(part.fieldB),
-              )
-            ? 'Pick a field from each query to join on'
-            : null;
-    const isIncomplete = setupStep !== null;
-
-    const blockingReason =
-        setupStep ??
-        (joinKeyErrors.length > 0
-            ? 'These queries cannot be joined on that field'
-            : unaccountedTotal > 0 && pivotSide === null
-              ? 'Too many extra fields to merge safely'
-              : pivotSide !== null && effectivePivotValues.length === 0
-                ? `Choose which ${pivotFieldLabel} values become columns`
-                : null);
-
-    const canRun =
-        completeParts.length > 0 &&
-        completeParts.length === effectiveParts.length &&
-        !!queryB.exploreName &&
-        queryB.metrics.length > 0 &&
-        joinKeyErrors.length === 0 &&
-        (unaccountedTotal === 0 ||
-            (pivotSide !== null && effectivePivotValues.length > 0));
-
-    const handleRun = useCallback(() => {
-        if (mergeQuery) run?.(mergeQuery);
-    }, [mergeQuery, run]);
+        joinKeyErrors,
+        isIncomplete,
+        blockingReason,
+        canRun,
+        handleRun,
+    } = useMergeSetup();
 
     const mergeError = mergeResults?.results.error ?? null;
 
@@ -645,19 +396,6 @@ export const MergeQueryStrip: FC = () => {
                         {blockingReason}
                     </Text>
                 )}
-                <Tooltip
-                    label={blockingReason ?? 'Compile and run the merged query'}
-                >
-                    <Button
-                        size="compact-sm"
-                        ml={blockingReason ? undefined : 'auto'}
-                        onClick={handleRun}
-                        loading={!!isRunning}
-                        disabled={!canRun}
-                    >
-                        Run merge
-                    </Button>
-                </Tooltip>
             </Group>
 
             {unaccountedTotal > 0 && pivotSide === null && !isIncomplete && (
